Refactor: change location for folder and i18n message
