+++ lang/en/auth.php
@@ -16,9 +16,4 @@ return [
     'failed' => 'These credentials do not match our records.',
     'password' => 'The provided password is incorrect.',
     'throttle' => 'Too many login attempts. Please try again in :seconds seconds.',
-    'unauthorized' => 'You do not have permission to access this resource.',
-    'login_success' => 'Login successful!',
-    'login_server_error' => 'An error occurred while trying to log in. Please try again later.',
-    'register_success' => 'Registration successful!',
-    'register_server_error' => 'An error occurred while trying to register. Please try again later.',
 ];
