--- lang/en/httpStatusMessage.php
@@ -1,5 +1,0 @@
-<?php
-
-return [
-    'not_found' => 'Not found.',
-];
