+++ lang/en/messages.php
@@ -1,0 +1,27 @@
+<?php
+
+return [
+    'users' => [
+        'get_all_success' => 'Users retrieved successfully.',
+        'get_all_fail' => 'Failed to retrieve users.',
+        'get_success' => 'User retrieved successfully.',
+        'get_fail' => 'Failed to retrieve user.',
+        'create_success' => 'User created successfully.',
+        'create_fail' => 'Failed to create user.',
+        'update_success' => 'User updated successfully.',
+        'update_fail' => 'Failed to update user.',
+        'delete_success' => 'User deleted successfully.',
+        'delete_fail' => 'Failed to delete user.',
+    ],
+    'auth' => [
+        'failed' => 'These credentials do not match our records.',
+        'unauthorized' => 'You do not have permission to access this resource.',
+        'login_success' => 'Login successful!',
+        'login_server_error' => 'An error occurred while trying to log in. Please try again later.',
+        'register_success' => 'Registration successful!',
+        'register_server_error' => 'An error occurred while trying to register. Please try again later.',
+    ],
+    'http_status_message' => [
+        'not_found' => 'Not found.'
+    ]
+];
