--- lang/en/users.php
@@ -1,14 +1,0 @@
-<?php
-
-return [
-    'get_all_success' => 'Users retrieved successfully.',
-    'get_all_fail' => 'Failed to retrieve users.',
-    'get_success' => 'User retrieved successfully.',
-    'get_fail' => 'Failed to retrieve user.',
-    'create_success' => 'User created successfully.',
-    'create_fail' => 'Failed to create user.',
-    'update_success' => 'User updated successfully.',
-    'update_fail' => 'Failed to update user.',
-    'delete_success' => 'User deleted successfully.',
-    'delete_fail' => 'Failed to delete user.',
-];
